pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "social-media-analytics"
version = "0.1.0"
description = "A modular analytics toolkit for Instagram & TikTok data from Apify"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
apify-client = "^1.6.0"
pandas = "^2.2.0"
numpy = "^1.26.0"
click = "^8.1.0"
pydantic = "^2.5.0"
pyarrow = "^15.0.0"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
rich = "^13.7.0"
scipy = "^1.11.0"
python-dotenv = "^1.1.1"
reportlab = "^4.4.2"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12"
types-python-dateutil = "^2.8.19"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "A", "COM", "C4", "DTZ", "ISC", "PIE", "PT", "RSE", "RET", "SIM", "TID", "PD", "PLE", "PLW"]
ignore = ["ANN101", "ANN102", "ANN401", "B008"]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
namespace_packages = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "if TYPE_CHECKING:",
]